requirements.txt
graphene-django==2.2.0
django==2.2.1
django-filter==2.1.0
django-neomodel==0.0.4
django-extensions==2.1.6
graphene-neo4j==2.2.3.6